Refactored the UI changes of MyRecipeDetailsScreen.kt by adding RecipeInfo function in CommonUi.kt. RecipeInfo shows the entire information of a recipe
